Migrating to Cypress 12
