+++ package.json
@@ -47,7 +47,7 @@
     "@typescript-eslint/parser": "^5.4.0",
     "axios": "^0.24.0",
     "cors": "^2.8.5",
-    "cypress": "^9.3.1",
+    "cypress": "12.17.2",
     "cypress-file-upload": "^5.0.8",
     "dotenv": "^16.0.0",
     "eslint": "^7.11.0",
